Two Stage Likelihood Ratio Test for Detection of Gene Expression Mid-parent Heterosis

This function apply two stage likelihood ratio test (2sLRT) to multi-family RNA-seq data with formats: three count matrices for famale, male and hybrid with genes in rows, samples in columns.

twoStageLRT(
  hybrid_mat,
  female_mat,
  male_mat,
  disp_mat = NULL,
  normalization = FALSE,
  norm_factors = NULL,
  n_fam_thres = 20,
  gene_groups_no = 3
)

Arguments

hybrid_mat

hybrids expression count with genes in rows, samples in columns.

female_mat

female expression count with genes in rows, samples in columns.

male_mat

male expression count with genes in rows, samples in columns.

disp_mat

if "Pois", use Pois LRT and only output Pois results; if NULL, use dispersion estimation output will have both NBLRT and PoisLRT results; else, use provided disp_mat for NB test, output will have both NBLRT and PoisLRT results.

normalization

logical, whether or not need normalization for the raw data matrices, if FALSE, treat the normalization factors all 1.

norm_factors

if normalization = TRUE and norm_factors = NULL, use DESeq normalization factors; if normalization = TRUE and norm_factors is a list contains three vectors female, male and hybrid normalization factors, use the provided norm_factors in tests.

n_fam_thres

genes with at least n_fam_thres null families will be used for dispersion estimation, otherwise, will use loess prediction to get dispersion, default is 20.

gene_groups_no

the number of clusters for genes, default is 3.

Value

if disp_mat = "Pois", return Poisson LRT for heterosis pvalue and adjusted pvalue from BH adjustment; if disp_mat = NULL or provided, return both NB and Poisson LRT for heterosis pvalue and qvalue and genewise dispersion, and the estimation of mu_gi, which is the female/male/hybrid_mat_normed.

Examples

library(TwoStageLRT)
data(sim_data)
res <- twoStageLRT(female_mat = sim_data$female_mat, male_mat = sim_data$male_mat,
  hybrid_mat = sim_data$hybrid_mat, disp_mat = NULL,
  normalization = TRUE, norm_factors = NULL,
  n_fam_thres = 10, gene_groups_no = 2)
#> Use DESeq estimated normalization factors. 
#> Gene clusters size: 
#> jacc_clust
#>  1  2 
#> 38  6 
#> Dispersion estimation for cluster 1 the number of families is 10 , number of genes is  38  
#> Dispersion estimation for cluster 2 the number of families is 14 , number of genes is  6  
#> Dispersion estimation summary:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
#> 0.02848 0.07723 0.09988 0.09786 0.11958 0.20517      56 
#> Output is from NB LRT for heterosis. 
summary(res$pval_NB)
#>     Family1            Family2             Family3            Family4        
#>  Min.   :0.000000   Min.   :0.0000001   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.003141   1st Qu.:0.0069335   1st Qu.:0.002582   1st Qu.:0.002255  
#>  Median :0.118183   Median :0.1741760   Median :0.171171   Median :0.068823  
#>  Mean   :0.270915   Mean   :0.3276612   Mean   :0.307413   Mean   :0.289863  
#>  3rd Qu.:0.485476   3rd Qu.:0.7213905   3rd Qu.:0.605963   3rd Qu.:0.546911  
#>  Max.   :0.986095   Max.   :0.9976220   Max.   :0.983298   Max.   :0.980386  
#>     Family5             Family6             Family7         
#>  Min.   :0.0000267   Min.   :0.0000025   Min.   :0.0000003  
#>  1st Qu.:0.0068711   1st Qu.:0.0166349   1st Qu.:0.0271803  
#>  Median :0.1875332   Median :0.2238102   Median :0.2298479  
#>  Mean   :0.2825152   Mean   :0.3198278   Mean   :0.3358626  
#>  3rd Qu.:0.4688109   3rd Qu.:0.6116209   3rd Qu.:0.6175672  
#>  Max.   :0.9787175   Max.   :0.9944113   Max.   :0.9823406  
#>     Family8             Family9             Family10           Family11        
#>  Min.   :0.0000001   Min.   :0.0000007   Min.   :0.000000   Min.   :0.0000178  
#>  1st Qu.:0.0119162   1st Qu.:0.0104258   1st Qu.:0.005567   1st Qu.:0.0045634  
#>  Median :0.1420970   Median :0.1877236   Median :0.207660   Median :0.1021768  
#>  Mean   :0.2831014   Mean   :0.3121191   Mean   :0.311800   Mean   :0.2469520  
#>  3rd Qu.:0.4754224   3rd Qu.:0.5799164   3rd Qu.:0.589639   3rd Qu.:0.4684446  
#>  Max.   :0.9886756   Max.   :0.9847935   Max.   :0.981980   Max.   :0.9951846  
#>     Family12            Family13            Family14           Family15       
#>  Min.   :0.0000033   Min.   :0.0000581   Min.   :0.000000   Min.   :0.000011  
#>  1st Qu.:0.0023896   1st Qu.:0.0062760   1st Qu.:0.004245   1st Qu.:0.013676  
#>  Median :0.1688473   Median :0.1519422   Median :0.133705   Median :0.127757  
#>  Mean   :0.3019228   Mean   :0.2775630   Mean   :0.266015   Mean   :0.274410  
#>  3rd Qu.:0.5840679   3rd Qu.:0.5041395   3rd Qu.:0.415331   3rd Qu.:0.508783  
#>  Max.   :0.9830689   Max.   :0.9602202   Max.   :0.998561   Max.   :0.984899  
#>     Family16            Family17          Family18           Family19        
#>  Min.   :0.0000005   Min.   :0.00000   Min.   :0.000000   Min.   :0.0000002  
#>  1st Qu.:0.0061767   1st Qu.:0.01028   1st Qu.:0.004807   1st Qu.:0.0069480  
#>  Median :0.1395831   Median :0.09403   Median :0.118840   Median :0.1133325  
#>  Mean   :0.3130438   Mean   :0.29644   Mean   :0.260671   Mean   :0.2432305  
#>  3rd Qu.:0.6443395   3rd Qu.:0.57196   3rd Qu.:0.475911   3rd Qu.:0.3882878  
#>  Max.   :0.9707993   Max.   :0.98860   Max.   :0.976484   Max.   :0.9961769  
#>     Family20            Family21           Family22            Family23       
#>  Min.   :0.0000008   Min.   :0.000000   Min.   :0.0000008   Min.   :0.000000  
#>  1st Qu.:0.0049227   1st Qu.:0.004765   1st Qu.:0.0117947   1st Qu.:0.003676  
#>  Median :0.0942494   Median :0.108498   Median :0.1265363   Median :0.148339  
#>  Mean   :0.2628955   Mean   :0.287513   Mean   :0.3112448   Mean   :0.298463  
#>  3rd Qu.:0.5078960   3rd Qu.:0.529812   3rd Qu.:0.5314887   3rd Qu.:0.632580  
#>  Max.   :0.9504680   Max.   :0.978376   Max.   :0.9987275   Max.   :0.936341  
#>     Family24            Family25            Family26          Family27       
#>  Min.   :0.0000001   Min.   :0.0000012   Min.   :0.00000   Min.   :0.000003  
#>  1st Qu.:0.0083047   1st Qu.:0.0024654   1st Qu.:0.01154   1st Qu.:0.004318  
#>  Median :0.0953822   Median :0.1919833   Median :0.10064   Median :0.117682  
#>  Mean   :0.2239814   Mean   :0.2846438   Mean   :0.26566   Mean   :0.258568  
#>  3rd Qu.:0.3657227   3rd Qu.:0.4949094   3rd Qu.:0.48037   3rd Qu.:0.468698  
#>  Max.   :0.8978113   Max.   :0.9974654   Max.   :0.99950   Max.   :0.985229  
#>     Family28            Family29           Family30           Family31        
#>  Min.   :0.0000005   Min.   :0.000000   Min.   :0.000000   Min.   :0.0000016  
#>  1st Qu.:0.0167739   1st Qu.:0.002964   1st Qu.:0.003151   1st Qu.:0.0154260  
#>  Median :0.1785930   Median :0.060949   Median :0.094312   Median :0.1771076  
#>  Mean   :0.2959248   Mean   :0.260574   Mean   :0.277829   Mean   :0.3027976  
#>  3rd Qu.:0.4969052   3rd Qu.:0.537297   3rd Qu.:0.599861   3rd Qu.:0.6240096  
#>  Max.   :0.9986734   Max.   :0.934492   Max.   :0.990961   Max.   :0.9937906  
#>     Family32          Family33            Family34            Family35        
#>  Min.   :0.00000   Min.   :0.0000001   Min.   :0.0000027   Min.   :0.0000099  
#>  1st Qu.:0.01098   1st Qu.:0.0185540   1st Qu.:0.0040176   1st Qu.:0.0059170  
#>  Median :0.17624   Median :0.1686602   Median :0.1302135   Median :0.1213127  
#>  Mean   :0.31275   Mean   :0.3060778   Mean   :0.3173386   Mean   :0.2696730  
#>  3rd Qu.:0.52803   3rd Qu.:0.6171712   3rd Qu.:0.5816214   3rd Qu.:0.5119552  
#>  Max.   :0.99813   Max.   :0.9911531   Max.   :0.9948841   Max.   :0.9417651  
#>     Family36            Family37          Family38          Family39       
#>  Min.   :0.0000002   Min.   :0.00000   Min.   :0.00000   Min.   :0.000000  
#>  1st Qu.:0.0024030   1st Qu.:0.02541   1st Qu.:0.01368   1st Qu.:0.009635  
#>  Median :0.0810233   Median :0.36893   Median :0.09693   Median :0.109436  
#>  Mean   :0.2900108   Mean   :0.38130   Mean   :0.27483   Mean   :0.290681  
#>  3rd Qu.:0.6087278   3rd Qu.:0.68574   3rd Qu.:0.54243   3rd Qu.:0.564344  
#>  Max.   :0.9990351   Max.   :0.99889   Max.   :0.99670   Max.   :0.992877  
#>     Family40            Family41          Family42            Family43        
#>  Min.   :0.0000002   Min.   :0.00000   Min.   :0.0000773   Min.   :0.0000265  
#>  1st Qu.:0.0024183   1st Qu.:0.01139   1st Qu.:0.0045812   1st Qu.:0.0033913  
#>  Median :0.1331527   Median :0.19650   Median :0.0729591   Median :0.0930332  
#>  Mean   :0.2804671   Mean   :0.29937   Mean   :0.2505627   Mean   :0.2485346  
#>  3rd Qu.:0.5741628   3rd Qu.:0.51950   3rd Qu.:0.4685585   3rd Qu.:0.4704084  
#>  Max.   :0.9922430   Max.   :0.99224   Max.   :0.9726509   Max.   :0.9781127  
#>     Family44          Family45            Family46            Family47        
#>  Min.   :0.00000   Min.   :0.0000005   Min.   :0.0000351   Min.   :0.0000002  
#>  1st Qu.:0.01192   1st Qu.:0.0090412   1st Qu.:0.0133569   1st Qu.:0.0082155  
#>  Median :0.13748   Median :0.1502580   Median :0.1824533   Median :0.1625227  
#>  Mean   :0.29312   Mean   :0.3103698   Mean   :0.3053191   Mean   :0.2873455  
#>  3rd Qu.:0.51442   3rd Qu.:0.5810211   3rd Qu.:0.5199578   3rd Qu.:0.6193643  
#>  Max.   :0.97861   Max.   :0.9990539   Max.   :0.9991870   Max.   :0.9946073  
#>     Family48           Family49           Family50           Family51        
#>  Min.   :0.000000   Min.   :0.000002   Min.   :0.000000   Min.   :0.0000003  
#>  1st Qu.:0.002824   1st Qu.:0.020037   1st Qu.:0.004531   1st Qu.:0.0033055  
#>  Median :0.124718   Median :0.191693   Median :0.262633   Median :0.1422642  
#>  Mean   :0.284637   Mean   :0.300176   Mean   :0.299381   Mean   :0.2789219  
#>  3rd Qu.:0.544744   3rd Qu.:0.535393   3rd Qu.:0.549738   3rd Qu.:0.5374044  
#>  Max.   :0.948074   Max.   :0.985514   Max.   :0.998906   Max.   :0.9643608  
#>     Family52            Family53            Family54          Family55       
#>  Min.   :0.0000317   Min.   :0.0000007   Min.   :0.00000   Min.   :0.000000  
#>  1st Qu.:0.0066420   1st Qu.:0.0066544   1st Qu.:0.01219   1st Qu.:0.006732  
#>  Median :0.1352907   Median :0.0477760   Median :0.14506   Median :0.119319  
#>  Mean   :0.2978497   Mean   :0.2297590   Mean   :0.28206   Mean   :0.271453  
#>  3rd Qu.:0.5426673   3rd Qu.:0.4187315   3rd Qu.:0.49639   3rd Qu.:0.521014  
#>  Max.   :0.9679563   Max.   :0.9675183   Max.   :0.97144   Max.   :0.979332  
#>     Family56           Family57            Family58            Family59        
#>  Min.   :0.000000   Min.   :0.0000003   Min.   :0.0000001   Min.   :0.0000001  
#>  1st Qu.:0.009565   1st Qu.:0.0031973   1st Qu.:0.0038704   1st Qu.:0.0053069  
#>  Median :0.247648   Median :0.1798209   Median :0.0846215   Median :0.1614354  
#>  Mean   :0.302916   Mean   :0.3032605   Mean   :0.2431680   Mean   :0.3295925  
#>  3rd Qu.:0.553829   3rd Qu.:0.5686231   3rd Qu.:0.4552893   3rd Qu.:0.6387586  
#>  Max.   :0.984883   Max.   :0.9692497   Max.   :0.9999417   Max.   :0.9928302  
#>     Family60            Family61            Family62        
#>  Min.   :0.0000001   Min.   :0.0000161   Min.   :0.0000006  
#>  1st Qu.:0.0067160   1st Qu.:0.0095365   1st Qu.:0.0149339  
#>  Median :0.1578514   Median :0.1149514   Median :0.1551578  
#>  Mean   :0.2912868   Mean   :0.2397425   Mean   :0.3057115  
#>  3rd Qu.:0.5179853   3rd Qu.:0.4196334   3rd Qu.:0.6020251  
#>  Max.   :0.9807262   Max.   :0.9783788   Max.   :0.9806566  
#>     Family63            Family64            Family65            Family66       
#>  Min.   :0.0000002   Min.   :0.0000001   Min.   :0.0000012   Min.   :0.000000  
#>  1st Qu.:0.0023005   1st Qu.:0.0126765   1st Qu.:0.0159003   1st Qu.:0.008235  
#>  Median :0.1503974   Median :0.2627176   Median :0.1555779   Median :0.103779  
#>  Mean   :0.2773244   Mean   :0.3572594   Mean   :0.2998776   Mean   :0.273859  
#>  3rd Qu.:0.4706248   3rd Qu.:0.6434151   3rd Qu.:0.5807417   3rd Qu.:0.487287  
#>  Max.   :0.9761202   Max.   :0.9913016   Max.   :0.9845407   Max.   :0.976124  
#>     Family67            Family68          Family69            Family70        
#>  Min.   :0.0000001   Min.   :0.00000   Min.   :0.0000015   Min.   :0.0000016  
#>  1st Qu.:0.0066113   1st Qu.:0.00932   1st Qu.:0.0057372   1st Qu.:0.0083321  
#>  Median :0.0891337   Median :0.16931   Median :0.0813863   Median :0.1721872  
#>  Mean   :0.2571368   Mean   :0.30869   Mean   :0.2718651   Mean   :0.3047718  
#>  3rd Qu.:0.4828021   3rd Qu.:0.63087   3rd Qu.:0.5574875   3rd Qu.:0.5250765  
#>  Max.   :0.9355606   Max.   :0.99908   Max.   :0.9512293   Max.   :0.9878447  
#>     Family71           Family72            Family73            Family74       
#>  Min.   :0.000000   Min.   :0.0000005   Min.   :0.0000002   Min.   :0.000000  
#>  1st Qu.:0.007622   1st Qu.:0.0042383   1st Qu.:0.0038049   1st Qu.:0.006439  
#>  Median :0.176379   Median :0.1460670   Median :0.1085119   Median :0.107909  
#>  Mean   :0.275988   Mean   :0.2839252   Mean   :0.2522503   Mean   :0.265350  
#>  3rd Qu.:0.514795   3rd Qu.:0.4935596   3rd Qu.:0.4445520   3rd Qu.:0.455792  
#>  Max.   :0.971724   Max.   :0.9988509   Max.   :0.9929457   Max.   :0.981676  
#>     Family75            Family76           Family77           Family78       
#>  Min.   :0.0000006   Min.   :0.000000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.0044645   1st Qu.:0.009659   1st Qu.:0.005268   1st Qu.:0.008821  
#>  Median :0.1692660   Median :0.117087   Median :0.106048   Median :0.168593  
#>  Mean   :0.3262287   Mean   :0.300347   Mean   :0.272086   Mean   :0.298540  
#>  3rd Qu.:0.6784842   3rd Qu.:0.617577   3rd Qu.:0.530501   3rd Qu.:0.574488  
#>  Max.   :0.9991996   Max.   :0.977342   Max.   :0.997269   Max.   :0.970074  
#>     Family79            Family80            Family81           Family82       
#>  Min.   :0.0000038   Min.   :0.0000002   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.0039837   1st Qu.:0.0022249   1st Qu.:0.005947   1st Qu.:0.003317  
#>  Median :0.2432629   Median :0.1646977   Median :0.062574   Median :0.096183  
#>  Mean   :0.2988813   Mean   :0.2886543   Mean   :0.250991   Mean   :0.312053  
#>  3rd Qu.:0.5527451   3rd Qu.:0.5252224   3rd Qu.:0.462126   3rd Qu.:0.652981  
#>  Max.   :0.9782831   Max.   :0.9947953   Max.   :0.941262   Max.   :0.999411  
#>     Family83           Family84            Family85           Family86       
#>  Min.   :0.000000   Min.   :0.0000113   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.005788   1st Qu.:0.0154479   1st Qu.:0.004237   1st Qu.:0.008897  
#>  Median :0.252953   Median :0.1906126   Median :0.193249   Median :0.141271  
#>  Mean   :0.342483   Mean   :0.2904036   Mean   :0.274805   Mean   :0.301919  
#>  3rd Qu.:0.652039   3rd Qu.:0.5204179   3rd Qu.:0.492677   3rd Qu.:0.608218  
#>  Max.   :0.960439   Max.   :0.9641852   Max.   :0.986025   Max.   :0.989554  
#>     Family87            Family88            Family89            Family90       
#>  Min.   :0.0000007   Min.   :0.0000001   Min.   :0.0000001   Min.   :0.000000  
#>  1st Qu.:0.0057953   1st Qu.:0.0059527   1st Qu.:0.0060745   1st Qu.:0.004772  
#>  Median :0.1437021   Median :0.0907612   Median :0.1429001   Median :0.186395  
#>  Mean   :0.3002562   Mean   :0.2426531   Mean   :0.2826259   Mean   :0.294279  
#>  3rd Qu.:0.5450359   3rd Qu.:0.4224134   3rd Qu.:0.5636609   3rd Qu.:0.560015  
#>  Max.   :0.9870277   Max.   :0.9789530   Max.   :0.9872134   Max.   :0.964583  
#>     Family91         Family92           Family93            Family94        
#>  Min.   :0.0000   Min.   :0.000000   Min.   :0.0000006   Min.   :0.0000007  
#>  1st Qu.:0.0136   1st Qu.:0.005689   1st Qu.:0.0034680   1st Qu.:0.0029844  
#>  Median :0.1780   Median :0.192139   Median :0.2076622   Median :0.1153720  
#>  Mean   :0.3072   Mean   :0.299734   Mean   :0.3003144   Mean   :0.2578489  
#>  3rd Qu.:0.5663   3rd Qu.:0.596648   3rd Qu.:0.5570118   3rd Qu.:0.5231216  
#>  Max.   :0.9871   Max.   :0.983718   Max.   :0.9970221   Max.   :0.9673605  
#>     Family95           Family96           Family97          Family98        
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.00000   Min.   :0.0000041  
#>  1st Qu.:0.003125   1st Qu.:0.006799   1st Qu.:0.01772   1st Qu.:0.0168081  
#>  Median :0.125581   Median :0.101626   Median :0.21532   Median :0.1417884  
#>  Mean   :0.295976   Mean   :0.261367   Mean   :0.32573   Mean   :0.2729182  
#>  3rd Qu.:0.580549   3rd Qu.:0.515920   3rd Qu.:0.61764   3rd Qu.:0.5042130  
#>  Max.   :0.991153   Max.   :0.995457   Max.   :0.98954   Max.   :0.9408215  
#>     Family99          Family100           Family101           Family102        
#>  Min.   :0.000000   Min.   :0.0000022   Min.   :0.0000001   Min.   :0.0000001  
#>  1st Qu.:0.007128   1st Qu.:0.0183956   1st Qu.:0.0043469   1st Qu.:0.0030543  
#>  Median :0.117145   Median :0.1372816   Median :0.0918666   Median :0.2127186  
#>  Mean   :0.277294   Mean   :0.2694521   Mean   :0.2804457   Mean   :0.3210234  
#>  3rd Qu.:0.520090   3rd Qu.:0.4670544   3rd Qu.:0.5671621   3rd Qu.:0.6420581  
#>  Max.   :0.991422   Max.   :0.9624432   Max.   :0.9914574   Max.   :0.9988530  
#>    Family103           Family104           Family105           Family106       
#>  Min.   :0.0000002   Min.   :0.0000005   Min.   :0.0000002   Min.   :0.000003  
#>  1st Qu.:0.0046776   1st Qu.:0.0049236   1st Qu.:0.0155700   1st Qu.:0.006651  
#>  Median :0.1615023   Median :0.1836819   Median :0.2284029   Median :0.246758  
#>  Mean   :0.2972274   Mean   :0.2774310   Mean   :0.3441116   Mean   :0.326803  
#>  3rd Qu.:0.5331151   3rd Qu.:0.4724934   3rd Qu.:0.6621673   3rd Qu.:0.609846  
#>  Max.   :0.9921426   Max.   :0.9799412   Max.   :0.9828324   Max.   :0.995812  
#>    Family107           Family108          Family109           Family110        
#>  Min.   :0.0000032   Min.   :0.000001   Min.   :0.0000032   Min.   :0.0000543  
#>  1st Qu.:0.0038181   1st Qu.:0.009318   1st Qu.:0.0149764   1st Qu.:0.0091751  
#>  Median :0.1549374   Median :0.127638   Median :0.0816055   Median :0.1480598  
#>  Mean   :0.2745862   Mean   :0.262713   Mean   :0.2651890   Mean   :0.3089273  
#>  3rd Qu.:0.5485569   3rd Qu.:0.468012   3rd Qu.:0.4882533   3rd Qu.:0.6162221  
#>  Max.   :0.9783487   Max.   :0.996855   Max.   :0.9754041   Max.   :0.9994839  
#>    Family111          Family112          Family113          Family114        
#>  Min.   :0.000003   Min.   :0.000000   Min.   :0.000000   Min.   :0.0000073  
#>  1st Qu.:0.005117   1st Qu.:0.003446   1st Qu.:0.009863   1st Qu.:0.0054214  
#>  Median :0.111583   Median :0.110739   Median :0.102419   Median :0.2029466  
#>  Mean   :0.280357   Mean   :0.278376   Mean   :0.266822   Mean   :0.3193826  
#>  3rd Qu.:0.507195   3rd Qu.:0.580344   3rd Qu.:0.465965   3rd Qu.:0.6481834  
#>  Max.   :0.992971   Max.   :0.989915   Max.   :0.997947   Max.   :0.9672457  
#>    Family115           Family116          Family117           Family118       
#>  Min.   :0.0000034   Min.   :0.000000   Min.   :0.0000105   Min.   :0.000000  
#>  1st Qu.:0.0063422   1st Qu.:0.003688   1st Qu.:0.0034312   1st Qu.:0.001296  
#>  Median :0.1468664   Median :0.092096   Median :0.1921112   Median :0.117453  
#>  Mean   :0.2965964   Mean   :0.263637   Mean   :0.3166161   Mean   :0.267357  
#>  3rd Qu.:0.5360831   3rd Qu.:0.492042   3rd Qu.:0.6055810   3rd Qu.:0.482176  
#>  Max.   :0.9913336   Max.   :0.996590   Max.   :0.9846471   Max.   :0.993374  
#>    Family119          Family120          Family121           Family122        
#>  Min.   :0.000000   Min.   :0.000002   Min.   :0.0000006   Min.   :0.0000032  
#>  1st Qu.:0.005059   1st Qu.:0.011321   1st Qu.:0.0052544   1st Qu.:0.0027887  
#>  Median :0.148445   Median :0.115797   Median :0.1743845   Median :0.1413960  
#>  Mean   :0.297388   Mean   :0.272448   Mean   :0.2892808   Mean   :0.3085612  
#>  3rd Qu.:0.572358   3rd Qu.:0.521841   3rd Qu.:0.5215983   3rd Qu.:0.6021803  
#>  Max.   :0.986147   Max.   :0.984711   Max.   :0.9611275   Max.   :0.9943156  
#>    Family123         Family124           Family125           Family126       
#>  Min.   :0.00000   Min.   :0.0000123   Min.   :0.0000004   Min.   :0.000000  
#>  1st Qu.:0.01079   1st Qu.:0.0021977   1st Qu.:0.0060740   1st Qu.:0.006251  
#>  Median :0.07018   Median :0.1197218   Median :0.0886303   Median :0.113359  
#>  Mean   :0.30033   Mean   :0.2651718   Mean   :0.2650141   Mean   :0.264032  
#>  3rd Qu.:0.57776   3rd Qu.:0.5085461   3rd Qu.:0.5136590   3rd Qu.:0.442524  
#>  Max.   :0.99399   Max.   :0.9789264   Max.   :0.9948257   Max.   :0.986820  
#>    Family127          Family128           Family129           Family130        
#>  Min.   :0.000000   Min.   :0.0000001   Min.   :0.0000091   Min.   :0.0000001  
#>  1st Qu.:0.003513   1st Qu.:0.0023866   1st Qu.:0.0130246   1st Qu.:0.0018425  
#>  Median :0.195310   Median :0.0964767   Median :0.1146152   Median :0.0683742  
#>  Mean   :0.293321   Mean   :0.2324637   Mean   :0.2669313   Mean   :0.2720158  
#>  3rd Qu.:0.501380   3rd Qu.:0.4153151   3rd Qu.:0.4755274   3rd Qu.:0.5025386  
#>  Max.   :0.999108   Max.   :0.8778796   Max.   :0.9944308   Max.   :0.9915402  
#>    Family131           Family132           Family133           Family134       
#>  Min.   :0.0000004   Min.   :0.0000008   Min.   :0.0000001   Min.   :0.000000  
#>  1st Qu.:0.0094666   1st Qu.:0.0179517   1st Qu.:0.0034862   1st Qu.:0.003371  
#>  Median :0.2028544   Median :0.1773721   Median :0.0730131   Median :0.162994  
#>  Mean   :0.3174130   Mean   :0.3182292   Mean   :0.2845872   Mean   :0.289595  
#>  3rd Qu.:0.6702761   3rd Qu.:0.5836910   3rd Qu.:0.5830014   3rd Qu.:0.500544  
#>  Max.   :0.9910362   Max.   :0.9906298   Max.   :0.9970272   Max.   :0.988354  
#>    Family135           Family136           Family137        
#>  Min.   :0.0000005   Min.   :0.0000001   Min.   :0.0000001  
#>  1st Qu.:0.0042483   1st Qu.:0.0042936   1st Qu.:0.0050699  
#>  Median :0.0978809   Median :0.0702976   Median :0.1372858  
#>  Mean   :0.2482284   Mean   :0.2628503   Mean   :0.2958454  
#>  3rd Qu.:0.4523456   3rd Qu.:0.4325040   3rd Qu.:0.5997733  
#>  Max.   :0.9772252   Max.   :0.9853099   Max.   :0.9927118  
#>    Family138           Family139           Family140        
#>  Min.   :0.0000002   Min.   :0.0000086   Min.   :0.0000008  
#>  1st Qu.:0.0028837   1st Qu.:0.0079294   1st Qu.:0.0169848  
#>  Median :0.0736912   Median :0.1524986   Median :0.1858294  
#>  Mean   :0.2764328   Mean   :0.2861767   Mean   :0.2892592  
#>  3rd Qu.:0.5398735   3rd Qu.:0.5031207   3rd Qu.:0.4657078  
#>  Max.   :0.9866726   Max.   :0.9587598   Max.   :0.9984249  
#>    Family141           Family142           Family143        
#>  Min.   :0.0000322   Min.   :0.0000038   Min.   :0.0000002  
#>  1st Qu.:0.0141654   1st Qu.:0.0133179   1st Qu.:0.0068828  
#>  Median :0.2598399   Median :0.2052282   Median :0.0653802  
#>  Mean   :0.3691653   Mean   :0.2899189   Mean   :0.2383843  
#>  3rd Qu.:0.7103046   3rd Qu.:0.4626940   3rd Qu.:0.4537916  
#>  Max.   :0.9975117   Max.   :0.9990482   Max.   :0.9971394  
#>    Family144           Family145           Family146           Family147      
#>  Min.   :0.0000004   Min.   :0.0000021   Min.   :0.0000025   Min.   :0.00000  
#>  1st Qu.:0.0054106   1st Qu.:0.0055401   1st Qu.:0.0086842   1st Qu.:0.00474  
#>  Median :0.1422316   Median :0.1080996   Median :0.1690145   Median :0.08619  
#>  Mean   :0.2856035   Mean   :0.2489457   Mean   :0.2928265   Mean   :0.25691  
#>  3rd Qu.:0.5465340   3rd Qu.:0.5048832   3rd Qu.:0.5218763   3rd Qu.:0.48819  
#>  Max.   :0.9815033   Max.   :0.9257502   Max.   :0.9845293   Max.   :0.98777  
#>    Family148           Family149           Family150       
#>  Min.   :0.0000015   Min.   :0.0000002   Min.   :0.000000  
#>  1st Qu.:0.0057573   1st Qu.:0.0207212   1st Qu.:0.005219  
#>  Median :0.1573198   Median :0.2197019   Median :0.084527  
#>  Mean   :0.3140863   Mean   :0.3175510   Mean   :0.259675  
#>  3rd Qu.:0.6021264   3rd Qu.:0.5587266   3rd Qu.:0.529593  
#>  Max.   :0.9835455   Max.   :0.9925152   Max.   :0.986592  
summary(res$pval_Pois)
#>     Family1             Family2           Family3           Family4        
#>  Min.   :0.0000000   Min.   :0.00000   Min.   :0.00000   Min.   :0.000000  
#>  1st Qu.:0.0000000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.000000  
#>  Median :0.0000000   Median :0.00000   Median :0.00000   Median :0.000000  
#>  Mean   :0.0441438   Mean   :0.09989   Mean   :0.08939   Mean   :0.072984  
#>  3rd Qu.:0.0005085   3rd Qu.:0.06180   3rd Qu.:0.00325   3rd Qu.:0.005733  
#>  Max.   :0.9160986   Max.   :0.98481   Max.   :0.94088   Max.   :0.885470  
#>     Family5             Family6            Family7           Family8         
#>  Min.   :0.0000000   Min.   :0.000000   Min.   :0.00000   Min.   :0.0000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.00000   1st Qu.:0.0000000  
#>  Median :0.0000000   Median :0.000000   Median :0.00000   Median :0.0000000  
#>  Mean   :0.0566968   Mean   :0.084482   Mean   :0.09687   Mean   :0.0815533  
#>  3rd Qu.:0.0000944   3rd Qu.:0.006558   3rd Qu.:0.01534   3rd Qu.:0.0001026  
#>  Max.   :0.8593717   Max.   :0.985665   Max.   :0.88056   Max.   :0.9165515  
#>     Family9            Family10           Family11            Family12       
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.000000   Median :0.0000000   Median :0.000000  
#>  Mean   :0.063480   Mean   :0.065854   Mean   :0.0514774   Mean   :0.079500  
#>  3rd Qu.:0.002306   3rd Qu.:0.002632   3rd Qu.:0.0009703   3rd Qu.:0.002624  
#>  Max.   :0.926136   Max.   :0.897796   Max.   :0.9727248   Max.   :0.912570  
#>     Family13           Family14            Family15          Family16      
#>  Min.   :0.000000   Min.   :0.0000000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.000000   Median :0.0000000   Median :0.00000   Median :0.00000  
#>  Mean   :0.052836   Mean   :0.0795880   Mean   :0.04789   Mean   :0.09434  
#>  3rd Qu.:0.003471   3rd Qu.:0.0001942   3rd Qu.:0.00126   3rd Qu.:0.01051  
#>  Max.   :0.834495   Max.   :0.9918518   Max.   :0.92332   Max.   :0.85078  
#>     Family17          Family18            Family19            Family20       
#>  Min.   :0.00000   Min.   :0.0000000   Min.   :0.0000000   Min.   :0.000000  
#>  1st Qu.:0.00000   1st Qu.:0.0000000   1st Qu.:0.0000000   1st Qu.:0.000000  
#>  Median :0.00000   Median :0.0000000   Median :0.0000000   Median :0.000000  
#>  Mean   :0.07438   Mean   :0.0457598   Mean   :0.0601420   Mean   :0.035403  
#>  3rd Qu.:0.00301   3rd Qu.:0.0002846   3rd Qu.:0.0000043   3rd Qu.:0.001649  
#>  Max.   :0.93720   Max.   :0.8590969   Max.   :0.9820672   Max.   :0.739628  
#>     Family21          Family22           Family23          Family24        
#>  Min.   :0.00000   Min.   :0.000000   Min.   :0.00000   Min.   :0.0000000  
#>  1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.00000   1st Qu.:0.0000000  
#>  Median :0.00000   Median :0.000000   Median :0.00000   Median :0.0000000  
#>  Mean   :0.06817   Mean   :0.100943   Mean   :0.07160   Mean   :0.0225338  
#>  3rd Qu.:0.00436   3rd Qu.:0.003917   3rd Qu.:0.01382   3rd Qu.:0.0000048  
#>  Max.   :0.90751   Max.   :0.995220   Max.   :0.73624   Max.   :0.3496974  
#>     Family25           Family26           Family27            Family28       
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.000000   Median :0.0000000   Median :0.000000  
#>  Mean   :0.073296   Mean   :0.070056   Mean   :0.0494986   Mean   :0.070414  
#>  3rd Qu.:0.002667   3rd Qu.:0.002725   3rd Qu.:0.0006147   3rd Qu.:0.002164  
#>  Max.   :0.985232   Max.   :0.997572   Max.   :0.9339595   Max.   :0.991490  
#>     Family29           Family30          Family31          Family32       
#>  Min.   :0.000000   Min.   :0.00000   Min.   :0.00000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.00000   Median :0.00000   Median :0.000000  
#>  Mean   :0.063074   Mean   :0.07164   Mean   :0.06971   Mean   :0.103150  
#>  3rd Qu.:0.003222   3rd Qu.:0.01071   3rd Qu.:0.01002   3rd Qu.:0.005627  
#>  Max.   :0.735409   Max.   :0.95124   Max.   :0.97305   Max.   :0.994037  
#>     Family33           Family34          Family35           Family36      
#>  Min.   :0.000000   Min.   :0.00000   Min.   :0.000000   Min.   :0.00000  
#>  1st Qu.:0.000000   1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.00000  
#>  Median :0.000000   Median :0.00000   Median :0.000000   Median :0.00000  
#>  Mean   :0.063983   Mean   :0.11167   Mean   :0.036108   Mean   :0.10617  
#>  3rd Qu.:0.006536   3rd Qu.:0.01566   3rd Qu.:0.001332   3rd Qu.:0.01275  
#>  Max.   :0.955971   Max.   :0.97231   Max.   :0.712427   Max.   :0.99430  
#>     Family37            Family38           Family39           Family40       
#>  Min.   :0.0000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000  
#>  Median :0.0000037   Median :0.000000   Median :0.000000   Median :0.000000  
#>  Mean   :0.0973989   Mean   :0.076947   Mean   :0.076387   Mean   :0.073122  
#>  3rd Qu.:0.0337244   3rd Qu.:0.003993   3rd Qu.:0.003638   3rd Qu.:0.006422  
#>  Max.   :0.9932346   Max.   :0.987511   Max.   :0.967391   Max.   :0.958957  
#>     Family41           Family42            Family43            Family44        
#>  Min.   :0.000000   Min.   :0.0000000   Min.   :0.0000000   Min.   :0.0000000  
#>  1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.0000000   1st Qu.:0.0000000  
#>  Median :0.000000   Median :0.0000000   Median :0.0000000   Median :0.0000000  
#>  Mean   :0.062238   Mean   :0.0512799   Mean   :0.0528415   Mean   :0.0794483  
#>  3rd Qu.:0.002857   3rd Qu.:0.0000305   3rd Qu.:0.0002615   3rd Qu.:0.0008105  
#>  Max.   :0.965899   Max.   :0.8833065   Max.   :0.8822219   Max.   :0.9032067  
#>     Family45           Family46           Family47           Family48       
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.000000   Median :0.000000   Median :0.000000  
#>  Mean   :0.085376   Mean   :0.080134   Mean   :0.070666   Mean   :0.061268  
#>  3rd Qu.:0.001577   3rd Qu.:0.002428   3rd Qu.:0.003354   3rd Qu.:0.001654  
#>  Max.   :0.995289   Max.   :0.996189   Max.   :0.972091   Max.   :0.750493  
#>     Family49           Family50           Family51           Family52       
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.000000   Median :0.000000   Median :0.000000  
#>  Mean   :0.070362   Mean   :0.050366   Mean   :0.038864   Mean   :0.074794  
#>  3rd Qu.:0.002193   3rd Qu.:0.001698   3rd Qu.:0.002078   3rd Qu.:0.006225  
#>  Max.   :0.912673   Max.   :0.993603   Max.   :0.792740   Max.   :0.842205  
#>     Family53            Family54           Family55           Family56       
#>  Min.   :0.0000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000  
#>  Median :0.0000000   Median :0.000000   Median :0.000000   Median :0.000000  
#>  Mean   :0.0493679   Mean   :0.050921   Mean   :0.056387   Mean   :0.051729  
#>  3rd Qu.:0.0000447   3rd Qu.:0.004657   3rd Qu.:0.001018   3rd Qu.:0.004595  
#>  Max.   :0.8701358   Max.   :0.907185   Max.   :0.882508   Max.   :0.935587  
#>     Family57           Family58            Family59          Family60      
#>  Min.   :0.000000   Min.   :0.0000000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.000000   Median :0.0000000   Median :0.00000   Median :0.00000  
#>  Mean   :0.066693   Mean   :0.0543721   Mean   :0.10178   Mean   :0.08290  
#>  3rd Qu.:0.003674   3rd Qu.:0.0001228   3rd Qu.:0.03911   3rd Qu.:0.00149  
#>  Max.   :0.838241   Max.   :0.9996840   Max.   :0.96565   Max.   :0.90353  
#>     Family61            Family62          Family63           Family64        
#>  Min.   :0.0000000   Min.   :0.00000   Min.   :0.000000   Min.   :0.0000000  
#>  1st Qu.:0.0000000   1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.0000000  
#>  Median :0.0000000   Median :0.00000   Median :0.000000   Median :0.0000003  
#>  Mean   :0.0440740   Mean   :0.06379   Mean   :0.067555   Mean   :0.0998742  
#>  3rd Qu.:0.0001376   3rd Qu.:0.00185   3rd Qu.:0.001267   3rd Qu.:0.0466200  
#>  Max.   :0.8778978   Max.   :0.89140   Max.   :0.894801   Max.   :0.9354869  
#>     Family65           Family66           Family67            Family68       
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.000000   Median :0.0000000   Median :0.000000  
#>  Mean   :0.062880   Mean   :0.078644   Mean   :0.0477565   Mean   :0.060501  
#>  3rd Qu.:0.003052   3rd Qu.:0.001605   3rd Qu.:0.0006157   3rd Qu.:0.005253  
#>  Max.   :0.922110   Max.   :0.889501   Max.   :0.7121470   Max.   :0.995248  
#>     Family69           Family70           Family71           Family72       
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.000000   Median :0.000000   Median :0.000000  
#>  Mean   :0.051011   Mean   :0.086683   Mean   :0.057146   Mean   :0.081255  
#>  3rd Qu.:0.001672   3rd Qu.:0.001991   3rd Qu.:0.003975   3rd Qu.:0.001748  
#>  Max.   :0.762721   Max.   :0.933275   Max.   :0.840380   Max.   :0.994962  
#>     Family73            Family74            Family75          Family76      
#>  Min.   :0.0000000   Min.   :0.0000000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.0000000   1st Qu.:0.0000000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.0000000   Median :0.0000000   Median :0.00000   Median :0.00000  
#>  Mean   :0.0487940   Mean   :0.0661452   Mean   :0.10169   Mean   :0.08211  
#>  3rd Qu.:0.0003608   3rd Qu.:0.0005492   3rd Qu.:0.03575   3rd Qu.:0.00560  
#>  Max.   :0.9603497   Max.   :0.8989022   Max.   :0.99497   Max.   :0.87219  
#>     Family77           Family78           Family79          Family80       
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.00000   Min.   :0.000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.00000   1st Qu.:0.000000  
#>  Median :0.000000   Median :0.000000   Median :0.00000   Median :0.000000  
#>  Mean   :0.050333   Mean   :0.085472   Mean   :0.06226   Mean   :0.049767  
#>  3rd Qu.:0.002012   3rd Qu.:0.006581   3rd Qu.:0.00549   3rd Qu.:0.002896  
#>  Max.   :0.989507   Max.   :0.871541   Max.   :0.89992   Max.   :0.980307  
#>     Family81            Family82         Family83            Family84       
#>  Min.   :0.0000000   Min.   :0.0000   Min.   :0.0000000   Min.   :0.000000  
#>  1st Qu.:0.0000000   1st Qu.:0.0000   1st Qu.:0.0000000   1st Qu.:0.000000  
#>  Median :0.0000000   Median :0.0000   Median :0.0000017   Median :0.000000  
#>  Mean   :0.0449665   Mean   :0.1080   Mean   :0.0899100   Mean   :0.053801  
#>  3rd Qu.:0.0000919   3rd Qu.:0.0174   3rd Qu.:0.0489565   3rd Qu.:0.002791  
#>  Max.   :0.7190105   Max.   :0.9976   Max.   :0.8107691   Max.   :0.922868  
#>     Family85            Family86           Family87           Family88        
#>  Min.   :0.0000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000  
#>  Median :0.0000000   Median :0.000000   Median :0.000000   Median :0.0000000  
#>  Mean   :0.0391190   Mean   :0.070328   Mean   :0.079127   Mean   :0.0433251  
#>  3rd Qu.:0.0004102   3rd Qu.:0.002676   3rd Qu.:0.009933   3rd Qu.:0.0002138  
#>  Max.   :0.9150409   Max.   :0.936312   Max.   :0.922575   Max.   :0.9110605  
#>     Family89          Family90           Family91           Family92      
#>  Min.   :0.00000   Min.   :0.000000   Min.   :0.000000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.000000   Median :0.000000   Median :0.00000  
#>  Mean   :0.04520   Mean   :0.049940   Mean   :0.067795   Mean   :0.06445  
#>  3rd Qu.:0.00257   3rd Qu.:0.003034   3rd Qu.:0.007091   3rd Qu.:0.01103  
#>  Max.   :0.94315   Max.   :0.793317   Max.   :0.922050   Max.   :0.92890  
#>     Family93          Family94           Family95           Family96        
#>  Min.   :0.00000   Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000  
#>  1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000  
#>  Median :0.00000   Median :0.000000   Median :0.000000   Median :0.0000000  
#>  Mean   :0.05284   Mean   :0.053869   Mean   :0.068919   Mean   :0.0556635  
#>  3rd Qu.:0.00220   3rd Qu.:0.002356   3rd Qu.:0.005252   3rd Qu.:0.0005714  
#>  Max.   :0.98480   Max.   :0.873758   Max.   :0.961330   Max.   :0.9733638  
#>     Family97          Family98           Family99           Family100        
#>  Min.   :0.00000   Min.   :0.000000   Min.   :0.0000000   Min.   :0.0000000  
#>  1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.0000000  
#>  Median :0.00000   Median :0.000000   Median :0.0000000   Median :0.0000000  
#>  Mean   :0.06883   Mean   :0.055211   Mean   :0.0563886   Mean   :0.0451257  
#>  3rd Qu.:0.01080   3rd Qu.:0.002453   3rd Qu.:0.0005049   3rd Qu.:0.0004811  
#>  Max.   :0.93638   Max.   :0.739545   Max.   :0.9468553   Max.   :0.7854173  
#>    Family101          Family102         Family103          Family104        
#>  Min.   :0.000000   Min.   :0.00000   Min.   :0.000000   Min.   :0.0000000  
#>  1st Qu.:0.000000   1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.0000000  
#>  Median :0.000000   Median :0.00000   Median :0.000000   Median :0.0000000  
#>  Mean   :0.073943   Mean   :0.07608   Mean   :0.091614   Mean   :0.0609285  
#>  3rd Qu.:0.002895   3rd Qu.:0.02444   3rd Qu.:0.005662   3rd Qu.:0.0003878  
#>  Max.   :0.955415   Max.   :0.99334   Max.   :0.963888   Max.   :0.9030445  
#>    Family105         Family106         Family107          Family108       
#>  Min.   :0.00000   Min.   :0.00000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.000000  
#>  Median :0.00000   Median :0.00000   Median :0.000000   Median :0.000000  
#>  Mean   :0.09198   Mean   :0.09528   Mean   :0.056475   Mean   :0.048930  
#>  3rd Qu.:0.02168   3rd Qu.:0.01000   3rd Qu.:0.002389   3rd Qu.:0.001031  
#>  Max.   :0.91826   Max.   :0.97268   Max.   :0.882111   Max.   :0.984299  
#>    Family109           Family110         Family111           Family112       
#>  Min.   :0.0000000   Min.   :0.00000   Min.   :0.0000000   Min.   :0.000000  
#>  1st Qu.:0.0000000   1st Qu.:0.00000   1st Qu.:0.0000000   1st Qu.:0.000000  
#>  Median :0.0000000   Median :0.00000   Median :0.0000000   Median :0.000000  
#>  Mean   :0.0658401   Mean   :0.07873   Mean   :0.0757839   Mean   :0.065134  
#>  3rd Qu.:0.0000761   3rd Qu.:0.01511   3rd Qu.:0.0001922   3rd Qu.:0.001643  
#>  Max.   :0.8853635   Max.   :0.99731   Max.   :0.9672294   Max.   :0.976296  
#>    Family113           Family114          Family115          Family116        
#>  Min.   :0.0000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000  
#>  Median :0.0000000   Median :0.000000   Median :0.000000   Median :0.0000000  
#>  Mean   :0.0652405   Mean   :0.073207   Mean   :0.080042   Mean   :0.0850080  
#>  3rd Qu.:0.0001496   3rd Qu.:0.009986   3rd Qu.:0.003213   3rd Qu.:0.0001154  
#>  Max.   :0.9887939   Max.   :0.817778   Max.   :0.955271   Max.   :0.9800718  
#>    Family117          Family118           Family119           Family120        
#>  Min.   :0.000000   Min.   :0.0000000   Min.   :0.0000000   Min.   :0.0000000  
#>  1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.0000000   1st Qu.:0.0000000  
#>  Median :0.000000   Median :0.0000000   Median :0.0000000   Median :0.0000000  
#>  Mean   :0.059498   Mean   :0.0729202   Mean   :0.0660673   Mean   :0.0539556  
#>  3rd Qu.:0.007504   3rd Qu.:0.0008786   3rd Qu.:0.0007892   3rd Qu.:0.0005803  
#>  Max.   :0.906033   Max.   :0.9662104   Max.   :0.9362989   Max.   :0.9233348  
#>    Family121           Family122          Family123          Family124        
#>  Min.   :0.0000000   Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000  
#>  Median :0.0000000   Median :0.000000   Median :0.000000   Median :0.0000000  
#>  Mean   :0.0507091   Mean   :0.095660   Mean   :0.076491   Mean   :0.0408773  
#>  3rd Qu.:0.0008452   3rd Qu.:0.005952   3rd Qu.:0.005956   3rd Qu.:0.0003485  
#>  Max.   :0.7368762   Max.   :0.972760   Max.   :0.969410   Max.   :0.8923366  
#>    Family125           Family126           Family127          Family128        
#>  Min.   :0.0000000   Min.   :0.0000000   Min.   :0.000000   Min.   :0.0000000  
#>  1st Qu.:0.0000000   1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.0000000  
#>  Median :0.0000000   Median :0.0000000   Median :0.000000   Median :0.0000000  
#>  Mean   :0.0580133   Mean   :0.0508139   Mean   :0.070823   Mean   :0.0254478  
#>  3rd Qu.:0.0000764   3rd Qu.:0.0000708   3rd Qu.:0.001158   3rd Qu.:0.0001182  
#>  Max.   :0.9762960   Max.   :0.9411081   Max.   :0.995562   Max.   :0.4215479  
#>    Family129           Family130          Family131         Family132       
#>  Min.   :0.0000000   Min.   :0.000000   Min.   :0.00000   Min.   :0.000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000   1st Qu.:0.00000   1st Qu.:0.000000  
#>  Median :0.0000000   Median :0.000000   Median :0.00000   Median :0.000000  
#>  Mean   :0.0664212   Mean   :0.097855   Mean   :0.07535   Mean   :0.094619  
#>  3rd Qu.:0.0001224   3rd Qu.:0.004958   3rd Qu.:0.02406   3rd Qu.:0.006784  
#>  Max.   :0.9705258   Max.   :0.964699   Max.   :0.96002   Max.   :0.948315  
#>    Family133          Family134          Family135           Family136        
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000   Min.   :0.0000000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.0000000  
#>  Median :0.000000   Median :0.000000   Median :0.0000000   Median :0.0000000  
#>  Mean   :0.073272   Mean   :0.081006   Mean   :0.0405508   Mean   :0.0889175  
#>  3rd Qu.:0.002906   3rd Qu.:0.004965   3rd Qu.:0.0004511   3rd Qu.:0.0001929  
#>  Max.   :0.983334   Max.   :0.931143   Max.   :0.8814490   Max.   :0.9160747  
#>    Family137         Family138          Family139          Family140       
#>  Min.   :0.00000   Min.   :0.000000   Min.   :0.000000   Min.   :0.000000  
#>  1st Qu.:0.00000   1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.000000  
#>  Median :0.00000   Median :0.000000   Median :0.000000   Median :0.000000  
#>  Mean   :0.08216   Mean   :0.073688   Mean   :0.057496   Mean   :0.052005  
#>  3rd Qu.:0.01568   3rd Qu.:0.007792   3rd Qu.:0.002347   3rd Qu.:0.001134  
#>  Max.   :0.96881   Max.   :0.930370   Max.   :0.817935   Max.   :0.991465  
#>    Family141         Family142           Family143           Family144      
#>  Min.   :0.00000   Min.   :0.0000000   Min.   :0.0000000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.0000000   1st Qu.:0.0000000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.0000000   Median :0.0000000   Median :0.00000  
#>  Mean   :0.08456   Mean   :0.0731342   Mean   :0.0393339   Mean   :0.08139  
#>  3rd Qu.:0.04872   3rd Qu.:0.0006185   3rd Qu.:0.0001028   3rd Qu.:0.00194  
#>  Max.   :0.98797   Max.   :0.9950044   Max.   :0.9840176   Max.   :0.89078  
#>    Family145          Family146          Family147           Family148      
#>  Min.   :0.000000   Min.   :0.000000   Min.   :0.0000000   Min.   :0.00000  
#>  1st Qu.:0.000000   1st Qu.:0.000000   1st Qu.:0.0000000   1st Qu.:0.00000  
#>  Median :0.000000   Median :0.000000   Median :0.0000000   Median :0.00000  
#>  Mean   :0.041084   Mean   :0.068947   Mean   :0.0337161   Mean   :0.07761  
#>  3rd Qu.:0.002024   3rd Qu.:0.001242   3rd Qu.:0.0006427   3rd Qu.:0.01504  
#>  Max.   :0.648778   Max.   :0.920392   Max.   :0.9336120   Max.   :0.89226  
#>    Family149           Family150       
#>  Min.   :0.0000000   Min.   :0.000000  
#>  1st Qu.:0.0000000   1st Qu.:0.000000  
#>  Median :0.0000002   Median :0.000000  
#>  Mean   :0.0704280   Mean   :0.055953  
#>  3rd Qu.:0.0034196   3rd Qu.:0.003099  
#>  Max.   :0.9673054   Max.   :0.945653  
summary(res$DISP)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> 0.02848 0.08031 0.09892 0.09567 0.11058 0.20517